Fix tests for < py3.6.2
